% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/smoking-data.R
\name{smoking_data}
\alias{smoking_data}
\title{Return a presentation-specific view of the smoking data}
\usage{
smoking_data(view = c("observational", "experimental", "mediation", "full"))
}
\arguments{
\item{view}{One of \code{"observational"}, \code{"experimental"}, \code{"mediation"},
or \code{"full"}.}
}
\value{
A tibble.
}
\description{
Return a presentation-specific view of the smoking data
}
