% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/palette.R
\name{uriah_palette}
\alias{uriah_palette}
\title{Uriah talks semantic color palette}
\usage{
uriah_palette()
}
\value{
A named character vector of semantic presentation colors.
}
\description{
Uriah talks semantic color palette
}
