% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/reactable.R
\name{uriah_reactable}
\alias{uriah_reactable}
\title{Shared reactable defaults for Uriah talks}
\usage{
uriah_reactable(data, ...)
}
\arguments{
\item{data}{A data frame.}

\item{...}{Additional arguments passed to \code{\link[reactable:reactable]{reactable::reactable()}}.}
}
\description{
Shared reactable defaults for Uriah talks
}
